refactor to isolate isce2 processing
