+++ src/hyp3_autorift/utils.py
@@ -122,90 +122,6 @@ def find_jpl_parameter_info(polygon: ogr.Geometry, parameter_file: str) -> dict:
     return parameter_info
 
 
-def format_tops_xml(reference, secondary, polarization, dem, orbits, xml_file='topsApp.xml'):
-    xml_template = f"""    <?xml version="1.0" encoding="UTF-8"?>
-    <topsApp>
-        <component name="topsinsar">
-            <component name="reference">
-                <property name="orbit directory">{orbits}</property>
-                <property name="auxiliary data directory">{orbits}</property>
-                <property name="output directory">reference</property>
-                <property name="safe">['{reference}.zip']</property>
-                <property name="polarization">{polarization}</property>
-            </component>
-            <component name="secondary">
-                <property name="orbit directory">{orbits}</property>
-                <property name="auxiliary data directory">{orbits}</property>
-                <property name="output directory">secondary</property>
-                <property name="safe">['{secondary}.zip']</property>
-                <property name="polarization">{polarization}</property>
-            </component>
-            <property name="demfilename">{dem}</property>
-            <property name="do interferogram">False</property>
-            <property name="do dense offsets">True</property>
-            <property name="do ESD">False</property>
-            <property name="do unwrap">False</property>
-            <property name="do unwrap 2 stage">False</property>
-            <property name="ampcor skip width">32</property>
-            <property name="ampcor skip height">32</property>
-            <property name="ampcor search window width">51</property>
-            <property name="ampcor search window height">51</property>
-            <property name="ampcor window width">32</property>
-            <property name="ampcor window height">32</property>
-        </component>
-    </topsApp>
-    """
-
-    with open(xml_file, 'w') as f:
-        f.write(textwrap.dedent(xml_template))
-
-
-class SysArgvManager:
-    """Context manager to clear and reset sys.argv
-
-    A bug in the ISCE2 Application class causes sys.argv to always be parsed when
-    no options are proved, even when setting `cmdline=[]`, preventing programmatic use.
-    """
-    def __init__(self):
-        self.argv = sys.argv.copy()
-
-    def __enter__(self):
-        sys.argv = sys.argv[:1]
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        sys.argv = self.argv
-
-
-def get_topsinsar_config():
-    from isce.applications.topsApp import TopsInSAR
-    with SysArgvManager():
-        insar = TopsInSAR(name="topsApp")
-        insar.configure()
-
-    config_data = {}
-    for name in ['reference', 'secondary']:
-        scene = insar.__getattribute__(name)
-
-        sensing_times = []
-        for swath in range(1, 4):
-            scene.configure()
-            scene.swathNumber = swath
-            scene.parse()
-            sensing_times.append(
-                (scene.product.sensingStart, scene.product.sensingStop)
-            )
-
-        sensing_start = min([sensing_time[0] for sensing_time in sensing_times])
-        sensing_stop = max([sensing_time[1] for sensing_time in sensing_times])
-
-        sensing_dt = (sensing_stop - sensing_start) / 2 + sensing_start
-
-        config_data[f'{name}_filename'] = Path(scene.safe[0]).name
-        config_data[f'{name}_dt'] = sensing_dt.strftime("%Y%m%dT%H:%M:%S.%f").rstrip('0')
-
-    return config_data
-
-
 def load_geospatial(infile: str, band: int = 1):
     ds = gdal.Open(infile, gdal.GA_ReadOnly)
